requirements.txt
Brlapi==0.5.5
Django==1.2.3
GnuPGInterface==0.3.2
Markdown==2.0.3
PAM==0.4.2
PIL==1.1.7
Pinax==0.9a1
Pygments==1.3.1
Twisted-Core==11.0.0
Twisted-Names==11.0.0
Twisted-Web==11.0.0
adium-theme-ubuntu==0.3.1
amqplib==1.0.2
anyjson==0.3.1
apt-xapian-index==0.44
apturl==0.5.1ubuntu1
ccsm==0.9.5.92
celery==2.5.1
chardet==2.0.1
command-not-found==0.2.44
compizconfig-python==0.9.5.94
configglue==1.0
creole==1.1
defer==1.0.2
diff-match-patch==20090804
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.16dev-r0
django-ajax-validation==0.1.3
django-announcements==0.1.2
django-atomformat==0.1.1.dev1
django-attachments==0.3.1
django-avatar==1.0.4
django-bookmarks==0.1.0
django-celery==2.5.1
django-debug-toolbar==0.8.3
django-email-confirmation==0.2.dev2
django-flag==0.2.dev9
django-friends==0.1.5
django-groups==0.1.dev10
django-imagekit==0.3.3
django-kombu==0.9.4
django-locations==0.1.1
django-mailer==0.2a1.dev1
django-markup==0.3
django-messages==0.4.2
django-microblogging==0.2.dev1
django-notification==0.1.4
django-oembed==0.1.1
django-openid==0.2.3.dev1
django-pagination==1.0.7
django-photologue==2.3
django-picklefield==0.2.0
django-sorting==0.1
django-staticfiles==0.2.0
django-swaps==0.5.3
django-tagging==0.3.1
django-tagging-ext==0.2.0
django-threadedcomments==0.5.2
django-timezones==0.2.dev1
django-typogrify==1.2.2
django-uni-form==0.7.0
django-voting==0.1
django-wakawaka==0.4.dev5
docutils==0.6
duplicity==0.6.15
foolscap==0.6.1
gdata==1.3.3
geopy==0.94
gnome-app-install==0.4.7-nmu1ubuntu2
httplib2==0.7.2
idios==0.2a1.dev2
ipython==0.10.2
jockey==0.9.4
keyring==0.6.2
kombu==2.1.2
launchpadlib==1.9.8
lazr.restfulclient==0.11.2
lazr.uri==1.0.2
louis==2.3.0
nvidia-common==0.0.0
oauth==1.0.1
oauth2==1.5.170
onboard==0.96.1
oneconf==0.2.6.7
papyon==0.5.5
pexpect==2.3
piston-mini-client==0.6
protobuf==2.4.0a
pyOpenSSL==0.12
pycrypto==2.3
pycups==1.9.59
pycurl==7.19.0
pyinotify==0.9.1
pyserial==2.5
pysmbc==1.0.10
python-apt==0.8.0ubuntu9
python-dateutil==1.4.1
python-debian==0.1.20ubuntu2
python-openid==2.2.5
python-twitter==0.5
python-virtkey==0.60.0
pytz==2010h
pyxdg==0.19
sessioninstaller==0.0.0
simplejson==2.0.9
smartypants==1.6.0.3
system-service==0.1.6
textile==2.1.4
ubuntu-sso-client==1.4.0
ubuntuone-couch==0.3.0
ubuntuone-installer==2.0.0
ubuntuone-storage-protocol==2.0.0
ufw==0.30.1-2ubuntu1
unattended-upgrades==0.1
usb-creator==0.2.23
virtualenv==1.6.1
virtualenvwrapper==2.7.1
vobject==0.8.1c
wadllib==1.2.0
wsgiref==0.1.2
wxPython==2.8.11.0
wxPython-common==2.8.11.0
xdiagnose==1.1
xkit==0.0.0
ybrowserauth==1.2
zope.interface==3.6.1
